code refactored into utils and main file
