+++ utils.py
@@ -1,0 +1,64 @@
+import matplotlib.pyplot as plt
+from sklearn.model_selection import train_test_split
+
+
+# Function for preprocessing (depends on the question)
+def preprocess_digits(dataset):
+    n_samples = len(dataset.images)
+    data = dataset.images.reshape((n_samples, -1))
+    label = dataset.target
+    return data, label
+
+# Function for data visulization (Use if want to see the maps)
+def data_viz(dataset):
+    # PART: sanity check visualization of the data
+    _, axes = plt.subplots(nrows=1, ncols=4, figsize=(10, 3))
+    for ax, image, label in zip(axes, dataset.images, dataset.target):
+        ax.set_axis_off()
+        ax.imshow(image, cmap=plt.cm.gray_r, interpolation="nearest")
+        ax.set_title("Training: %i" % label)
+
+# Generic Function for test,train & dev splits
+def train_dev_test_split(data, label, train_frac, dev_frac):
+    dev_test_frac = 1 - train_frac
+    x_train, x_dev_test, y_train, y_dev_test = train_test_split(
+        data, label, test_size=dev_test_frac, shuffle=True
+    )
+    x_test, x_dev, y_test, y_dev = train_test_split(
+        x_dev_test, y_dev_test, test_size=(dev_frac) / dev_test_frac, shuffle=True
+    )
+
+    return x_train, y_train, x_dev, y_dev, x_test, y_test
+
+
+# Main Function for hyper parameter tunning
+def h_param_tuning(h_param_comb, clf, x_train, y_train, x_dev, y_dev, metric):
+    best_metric = -1.0
+    best_model = None
+    best_h_params = None
+    # 2. For every combination-of-hyper-parameter values
+    for cur_h_params in h_param_comb:
+        # PART: setting up hyperparameter
+        hyper_params = cur_h_params
+        clf.set_params(**hyper_params)
+
+        # PART: Train model
+        # 2.a train the model
+        # Learn the digits on the train subset
+        clf.fit(x_train, y_train)
+
+        # print(cur_h_params)
+        # PART: get dev set predictions
+        predicted_dev = clf.predict(x_dev)
+
+        # 2.b compute the accuracy on the validation set
+        cur_metric = metric(y_pred=predicted_dev, y_true=y_dev)
+
+        # 3. identify the combination-of-hyper-parameter for which validation set accuracy is the highest.
+        if cur_metric > best_metric:
+            best_metric = cur_metric
+            best_model = clf
+            best_h_params = cur_h_params
+            print("Found new best metric with :" + str(cur_h_params))
+            print("New best val metric:" + str(cur_metric))
+    return best_model, best_metric, best_h_params
